Modale de suppression

+++ 13-beerpdomysql/brewery_list.php
@@ -43,7 +43,7 @@ $brands = $query->fetchAll();
                   <a href="./brewery_single.php?id=<?php echo $brand['id']?>" class="btn btn-info">+ d'infos</a>
                   <?php if(userIsLogged()) { ?>
                   <a href="./brewery_single.php?id=<?php echo $brand['id']?>" class="btn btn-warning">Modifier</a>
-                  <a href="./brewery_delete.php?id=<?php echo $brand['id']?>" class="btn btn-danger">Supprimer</a>
+                  <a href="./brewery_delete.php?id=<?php echo $brand['id']?>" class="btn btn-danger confirm-delete">Supprimer</a>
                   <?php } ?>
                   </td>
                 </tr>
